Module type Webdriver.S

type json = Yojson.Safe.t

type 'a io

The client I/O monad, within which communication happens with the WebDriver server.

Commands

type session

A connection to the WebDriver server. A session statefully holds the active windows, tabs, cookies, and other state required by the browser.

type 'a cmd = session:session -> 'a io

Every browser command takes place within a session.

module Infix : sig ... end

Since the ~session parameter is generally constant, this module provides a reader monad to sequence multiple commands within the same session. You can either pass explicitly the ~session argument or open this module.

module Error : sig ... end

All potentital errors raised by the WebDriver protocol.

exception Webdriver of Error.t

Every command that fails raises this exception, which contains some hint as to what went wrong.

Sessions

In order to create a session, a connection to a WebDriver-compatible browser must be established:

The host is a url where the WebDriver server can be accessed, for example "http://localhost:4444".

The capabilities describe the requested browser settings.

module Capabilities : sig ... end

The requested capabilities when creating a new session.

module Session : sig ... end

For creating and deleting sessions manually.

val run : host:string -> Capabilities.t -> 'a cmd -> 'a io

run ~host capabilities cmd is a helper function to create a new session on host with the required capabilities, execute the cmd within that session, and finally ensure that the session is deleted on termination of cmd (from its natural death or an exception.)

module Timeouts : sig ... end

Configure the timeouts for page loads, script execution and the implicit wait when searching for elements on a page.

Navigation

val goto : string -> unit cmd

goto url ask the browser to visit the page at url.

val current_url : string cmd

The current url.

val back : unit cmd

Click the browser back button, to reload the previous url in history.

val forward : unit cmd

Click the forward button, to move forward to the next url in history.

val refresh : unit cmd

Refresh the current url.

module Cookie : sig ... end

Cookies management.

HTML Elements

type elt

An HTML element.

type using = [

| `css
CSS selectors, like "h1 span"

| `link_text
The exact text in an <a>...</a>

| `partial_link_text
The partial text present in a link

| `tag_name
the HTML tag name, like "h1" or "div"

| `xpath
XPath query

]

A strategy to find an element on a page.

val find_first : ?⁠from:elt -> using -> string -> elt cmd

find_first `using "query" returns the first element that matches the query (interpreted with using). The element is searched inside the current frame of the current window, and if a ?from parent element is provided, the search takes place inside it.

raise (Webdriver { error = `no_such_element ; _ }) otherwise.

let* elt = find_first `css "h1 a" in ...
let* elt = find_first `xpath "//h1//a" in ...

val find_all : ?⁠from:elt -> using -> string -> elt list cmd

find_all `using "query" behaves like find_first, but returns a list of all elements matching the query.

val active : elt cmd

The currently active element on the page.

Inspecting HTML elements

val text : elt -> string cmd

The inner text of an element, ignoring any other markup.

val tag_name : elt -> string cmd

The HTML tag of an element, for example "div".

val attribute : elt -> string -> string cmd

attribute e attr returns the value of the HTML attribute attr of the element e.

val property : elt -> string -> string option cmd

property e prop returns Some value of the DOM property prop of the element e, or None if undefined.

val is_selected : elt -> bool cmd

The boolean status of a checkbox, a radio or an option in a select.

val is_enabled : elt -> bool cmd

The boolean status of an interactive element.

val css : elt -> string -> string cmd

css e prop returns the computed value of the css property prop for the element e.

val aria_role : elt -> string cmd

aria_role e returns the accessibility role of the element e. See ARIA Roles on MDN

val aria_label : elt -> json cmd

aria_role e returns the accessibility label of the element e. See ARIA Labels on MDN

type rect = {

x : int;

y : int;

width : int;

height : int;

}

val rect : elt -> rect cmd

The position and size of an element on the page.

Interactions

val click : elt -> unit cmd

Performs a mouse click on this element.

val clear : elt -> unit cmd

Clears the content of an input element.

val send_keys : elt -> string -> unit cmd

send_keys e str sends the string str to an input element, as if typed from a keyboard. For special keys like enter or backspace, use the predefined values in the module Key:

send_keys my_input ("hello" ^ Key.enter)

module Key : sig ... end

Special keys on a keyboard.

type action

A sequence of interactions from a user device.

val perform : action list -> unit cmd

perform actions executes the sequence of actions for each input source. The actions are synchronized vertically, such that:

perform [ mouse    [ `down button0 ; `pause  ; `up button0 ]
        ; keyboard [ `down "a"     ; `up "a" ; `pause      ]
        ]

The mouse starts a left click and the keyboard presses "a" simultaneously;

Then the keyboard releases the "a" key;

And finally the mouse releases the left click.

The `pause action does nothing and is used for synchronization.

The pressed keys and buttons stay pressed at the end of the interaction, unless explicitly released.

val release : unit cmd

release any pending keys or button from previous interactions.

val none : ?⁠name:string -> [ `pause ] list -> action

An inoperative device.

Keyboard actions

type key = [

| `pause
do nothing

| `down of Key.t
press down the key

| `up of Key.t
release a pressed key

]

An action from a keyboard.

val keyboard : ?⁠name:string -> key list -> action

keyboard keys is an action that

Pointer actions

type move = {

move_duration : int;
time taken by the move

move_origin : [ `viewport | `pointer | `elt of elt ];
relative position of move_x,move_y

move_x : int;

move_y : int;

}

A mouse/pointer movement. The target destination move_x,move_y is relative to the origin:

The `viewport uses the top left of the currently focused frame as the zero.

The `pointer uses the current position of the device on the screen.

The `elt e uses the top left position of the element e as the origin.

type button = int

An integer representing the nth button on a mouse.

val button0 : button

The left mouse button.

val button1 : button

The right mouse button.

type pointer = [

| `pause
do nothing

| `cancel
cancel the pointer current action

| `down of button
press down the button

| `up of button
release a pressed button

| `move of move
move the pointer

]

An action from a mouse/touch/pen device

val mouse : ?⁠name:string -> pointer list -> action

mouse actions describes the movement, click, etc of a mouse pointer.

val touch : ?⁠name:string -> pointer list -> action

touch actions describes the interactions of a touch finger device.

val pen : ?⁠name:string -> pointer list -> action

pen actions describes an interaction of a pencil.

Scroll wheel actions

type scroll = {

scroll_duration : int;
time taken by the scroll

scroll_origin : [ `viewport | `elt of elt ];
relative position of scroll_x,scroll_y

scroll_x : int;

scroll_y : int;

}

A scroll movement. The final scroll destination is relative to the origin.

type wheel = [

| `pause
do nothing

| `scroll of scroll
scroll to a position

]

An action from the scroll wheel.

val wheel : ?⁠name:string -> wheel list -> action

Document

val title : string cmd

The page title of the current document.

val source : string cmd

The HTML source code of the current document.

val print : string cmd

The current page, printed as a PDF in base64.

val screenshot : ?⁠elt:elt -> unit -> string cmd

Returns a screenshot of the current page, or of the provided ?elt. Encoded in base64.

val switch_to_frame : [ `top | `id of int | `elt of elt ] -> unit cmd

Focus the selected frame inside the current document.

The `top frame is the current document root.

The `id n frame is the nth frame in the page.

The `elt e frame is the frame associated with the HTML element e.

val switch_to_parent_frame : unit cmd

Focus the parent of the currently selected frame.

module Window : sig ... end

Windows and tabs management.

module Alert : sig ... end

Popup management: alert, confirm and prompt

Javascript execution

val execute : string -> json cmd

excute "js" runs the js on the current page, returning its result in json.

val execute_async : string -> json cmd

excute_async "js" runs the js asynchronously on the current page.