Module S.Timeouts

Configure the timeouts for page loads, script execution and the implicit wait when searching for elements on a page.

type t = {

script : [ `Never | `After of int ];
default is `After 30_000 ms (= 3s)

page_load : int;
default is 300_000 ms (= 300s)

implicit_wait : int;
default is 0 ms

}

val get : t cmd

get the current timeouts configuration.

val set : ?⁠script:[ `Never | `After of int ] -> ?⁠page_load:int -> ?⁠implicit_wait:int -> unit -> unit cmd

set ?script ?page_load ?implicit_wait () updates the configuration for the provided fields.