Module S.Session

For creating and deleting sessions manually.

val make : host:string -> Capabilities.t -> (session * json) io

make ~host cap creates a new session on host with the required capabilities. It returns the created session and a json describing the available capabilities of the session.

val delete : unit cmd

delete ~session closes the session and remove all of its windows, cookies, etc.